+++ pyqgiswps/owsutils/ows.py
@@ -1,6 +1,0 @@
-#
-# Indirection to owslib
-#
-
-from owslib.ows import BoundingBox # noqa F401
-