+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.6.6 - 2021-02-01
+
 * Install server in venv in docker image
 * Add support for inputs:
     * 'Mesh' layer parameter